pyproject.toml
[project]
name = "ripple"
version = "0.1.0"
description = "Multi-agent supply chain early warning intelligence"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langgraph>=0.2",
    "langsmith>=0.1",
    "langchain-anthropic>=0.3",
    "anthropic>=0.40",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "neo4j>=5.26",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "pymongo>=4.10",
    "langgraph-checkpoint-mongodb>=0.3",
    "redis>=5.2",
    "mem0ai[nlp]>=0.1",
    "feedparser>=6.0",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "pydantic>=2.10",
    "ollama>=0.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
